fix (Generator) Update To Layout Priority
This modification allows the user to specify a path they want their
uploader to be generated at.

The modification also allows the user not to specify a path if they are
generating the uploader in a Phoenix project >= 1.3. If running the
generator in a Phoenix project, the uploader will automatically be
generated in the appropriate folder.

If a path is not passed and the user is not generating the uploader in a
Phoenix project, then an exception will be raised.

* toucan *  11/96
                                 `| |
                    .-"-.__       | .'
                   /  ,-|  ```'-, || . `
                  /  /@)|_       `-,  |
                 /  ( ~ \_```""--.,_',  :
              _.'    \   /```''''""`^
             /        | /        `|   :
            /         \/          | |   .
           /      __/  |          |. |
          |     __/  / |          |     |
"-._      |    _/ _/=_//          ||  :
'.  `~~`~^|   /=/-_/_/`~~`~~~`~~`~^~`.
  `> ' .   \  ~/_/_ /"   `  .  ` ' .   |
 .' ,'~^~`^|   |~^`^~~`~~~^~~~^~`;       '
 .-'       | | |                  \ ` : `
           |  :|                  | : |
      jgs  |:  |                  ||    '
           | |/                   |  |   :
           |_/                    | .  '